pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "mrm-deepagent"
version = "0.1.0"
description = "Deep agent for generating and applying model risk management documentation updates."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "deepagents>=0.0.10",
  "langchain>=0.3.0",
  "langchain-google-genai>=2.1.0",
  "python-docx>=1.1.2",
  "pydantic>=2.8.0",
  "typer>=0.16.0",
  "rich>=14.0.0",
  "pyyaml>=6.0.2",
]

[project.scripts]
mrm-agent = "mrm_deepagent.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mrm_deepagent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/mrm_deepagent --cov-report=term-missing --cov-fail-under=90"
markers = ["live: marks tests requiring live Gemini API access"]
